Support long running item store tests
A page erase operation takes ~20ms of time. During this time no item can
be added to the item store.
Therefore if we want to insert a large number of items we need to
trigger the insert operation by a timer that gives enough time to insert
an item including erase and other pending operations on the system.

With the test method ItemStore_TimerAddItem() it becomes possible to
have a test that inserts many thousands of items in shorter time than it
would be done by the application.