Use hardware CRC block
CRC calculation is used for sensor data and device settings. Since the
device settings is a quite big data structure that needs to be checked,
the HW block should be used in order to be more efficient.
In this commit it is tested in the context of the sensor data.