Delete logged samples on POR
The measurement item log does not contain time information. Hence we do
not know how long the system is powered off when the battery is removed
and the interval between subsequently logged items may be considerable
different from the logging interval.
In order to avoid this situation all samples are removed on a POR.
A soft reset will not delete the available data points.
Deleting the flash is the easier solotion than just invalidating some
pages. This would require modification of the existing logic.