Avoid clash of flash erase and BLE startup
As the flash is a shared resource of the wireless stack processor and
the application processor, we need to take care that longrunning
operations on the flash do not interfere with any wireless operation.

The cleanup of measurement samples apparently was inhibiting the BLE
subsystem to start up. By postponing the item-store cleanup we can go
around this problem.