Fix clang-tidy warnings
Clang tidy is not fully reflecting the used compiler architecture. It is
issuing warnings where there are no other options.
One of these examples is the use of sprintf. Therefore we define an
implementation that is checked without warning.
Just having an empty definiton for the macros will lead to other
warnings in the code.