Introduce a battery level filter
The measured voltages show a high spread mainly with weak batteries.
This make the battery monitor as it was before very unreliable.
The introduce median filter stabilizes the application state and
remaining capacity considerable. If this is already enough has to be
figured out with extensive tests (also with weak batteries).
I fear that the dimension of the used vbat capacitor is not good enough
also to guarantee the required battery life.