Fix the structure of the advertisemen type
By moving the local name before the manufacturar data we can simply
reduce the length of both manufacturer-data and overall length in order
to hide the temperature and humidity.
When the overall lenght does not match the defined tags of th structure,
the function aci_update_adv_data returns an error.