Support debug pins PA12 and PB11
In order to optimize and debug the Demo-board firmware we need to be
able to non intrusively observe the application.
For this purpose two pins are made available on the demo board
connector.
With the new functionality these pins can be configured as output and
forced to a desired state. In this way the code can be instrumented
according to the debug or measurement scenario.