Fix log macros
In order to consistently replace all occorrencies of TraceMessage(..)
with a log macro, the log macro has to be corrected. We want to be abel
to write LOG_INFO("message") which was not possible with the versio as
it was before. With this little fix, the macros can be used with one or
more arguments.