Use log macros from Log.h header
In order to have a consistent output format one must not directly use
the function TraceMessage. Furthermore we do not want several versions
of the same log macro spread in various places.
In order to make the code compile with the global log macros, some
variables needed to be declared as [[maybe_unused]].