Improve the item store
When an item store is empty we should not erase its flash pages.
This requires a lot of current and uses up the erase cycles.

Erasing the empty item store happened when a the demo-board was booted
for the first time and also when users changed frequently the logging
intervall.