Change behavior of Presentation controller
The presentation controller does only react on the messages BLE ON/OFF
messages to show/hide the BLE symbol. It does not anymore evaluate the
Button press event since this would require to douplicate to much logic.
The logic to switch BLE ON/OFF should only be in one place.
Furthermore high contrast is enabled in case of application state
critical