Restrict write and erase area
Asserts are introduced in order to prevent damage from wrong write or
erase operations.
Since in some cases a flash erase address is read from flash, it might
happen, that the process gets a wrong address.
In these cases the programm must not damage the flash but just stay in
an inifinite loop. By removing the battery the errornous flash area
should be erased and the application can continue to work.