Fix Sensor-Controller
It showed that the sensor controller could get stuck in a state because
an I2c error condition was not received as expected.
With the provided fix this phenomenon is not seen anymore. A forced
shorting of i2c pins will lead to an unrecoverable error.
A reset of the device by removing the battery will be required to
recover from it.