Improve recovery from i2c errors
The I2c error behavior was still not good. The following improvements
have been added:
- Set the appropriate states in the i2c driver to initialize and release
  the driver.
- Check for failure when starting an i2c transaction.
- Allow to force a reset of the i2c peripheral upon errors.
- Add timouts to all susceptible states of the sensor controller
- include a dedicated error handling state in the sensor controller
With these changes we can recover from a short of SCL that is shorter
than the number of retries.
We still will endup in an assert state when the error condition persists
for longer time.