Prevent recovering from critical battery level
Havy loads on the battery such as download of large amount of logging
data may dramatically reduce the current voltage of the battery.
The less capacity the battery has, the more severe the voltage drop will
be.
As long as the voltage is above a critical level we allow the
application to recover and to adjust the application state when the
battery recovers.
But when the voltage drops below 2.5V we do not change the application
state anymore but assume that the battery needs to be changed and the
device needs to be reset.